Moved to Let's Encrypt (#51)
Our SSL certificate is expiring in 1 month. Instead of going through the process of SSL certificates one more time, we decided to try out Let's Encrypt. It was really easy to set it up via the CLI. The slow part was getting our tests and Vagrant to align. In this PR:

- Added Let's Encrypt documentation
- Updated Vagrant and Wercker to match Let's Encrypt filepaths
- Updated tests
- Drive-by addition of `pry` for debugging